agents/meal-agent/requirements.txt
# Strands Agents SDK
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# Bedrock AgentCore SDK
bedrock-agentcore>=0.1.0

# AWS SDK
boto3>=1.34.0

# HTTP requests for Spoonacular
httpx>=0.27.0

# Environment management
python-dotenv>=1.0.0

# Type hints
typing-extensions>=4.9.0
